Added fixes
