+++ homework_interfaces_8/src/index.ts
@@ -1,22 +1,78 @@
-class Address2 {
-    public constructor(public street: string, public city: string) {}
+import { Results, User } from './interfaces';
+async function fetchRandomUser(): Promise<Results> {
 
-    public getFullAddress(): string {
-        return `${this.street}, ${this.city}`;
+    try {
+        const response = await fetch('https://randomuser.me/api/');
+
+        if (!response.ok) {
+            throw new Error('Failed to fetch data');
+        }
+
+        const json = await response.json() as Results;
+
+        return json;
+    } catch (error) {
+        console.error('Error fetching random user:', error);
+        throw error;
     }
 }
 
-class Student2 {
-    public constructor(public name: string, public age: number, public address: Address2) {}
+class Obj2 {
+    public shortInfo: string;
+    public sumAge: number;
+    public fullLocation: string;
 
-    public introduce(): void {
-        console.log(`Hi, I'm ${this.name}, a ${this.age}-year-old student.`);
-    }
+    public constructor(user: User) {
 
-    public getDetails(): string {
-        return `${this.name}, Age: ${this.age}, Address: ${this.address.getFullAddress()}`;
+        this.shortInfo = `Name: ${user.name.first} ${user.name.last}, Gender: ${user.gender}, Email: ${user.email}`;
+        this.sumAge = user.dob.age;
+        this.fullLocation = `${user.location.city}, ${user.location.state}, ${user.location.country}`;
     }
 }
+
+(async () => {
+    const data = await fetchRandomUser();
+    const user = data.results[0];
+
+    const fullName = `${user.name.title} ${user.name.first} ${user.name.last}`;
+    console.log(`Full Name: ${fullName}`);
+
+    console.log(`Email: ${user.email}`);
+
+    const location = `${user.location.city}, ${user.location.state}, ${user.location.country}`;
+    console.log(`Location: ${location}`);
+
+    user.phone = '579-111-8250';
+    console.log(`Modified Phone Number: ${user.phone}`);
+
+    const obj2 = new Obj2(user);
+    console.log(obj2.shortInfo);
+    console.log(`Age: ${obj2.sumAge}`);
+    console.log(`Location: ${obj2.fullLocation}`);
+})();
+
+
+import { Book, Author, Publisher } from './class';
+
+const author = new Author('Paulo Coelho', 1947);
+const publisher = new Publisher('Harper Torch', 'Brazil');
+const book = new Book('The Alchemist', author, publisher, 1988);
+
+console.log(book.getBookInfo());
+console.log(`Author's Age: ${book.getAuthorAge(2025)}`);
+book.setPublishYear(2002);
+
+
+import { Address, Student } from './abstraction';
+
+const address = new Address('Main street 122', 'Kyiv');
+const student = new Student('Anna Kos', 19, address, 'Economical University');
+
+student.introduce();
+console.log(student.getDetails());
+
+
+import { Address2, Student2 } from './Address2';
 
 const newAddress = new Address2('Main St 123', 'Kyiv');
 const newStudent = new Student2('Anna', 19, newAddress);
